moved persistence to datastore
